plugin-server/requirements.txt
django>=5.1,<5.2
django-ninja>=1.3,<2.0
psycopg[binary]>=3.2,<4.0
pgvector>=0.3,<1.0
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.30,<1.0
whitenoise>=6.7,<7.0
django-unfold>=0.40,<1.0
mcp>=1.0,<2.0
